take same-size&place parents with you
when you leave.

sometimes, people like to stack elements exactly on top of one another, and dom
destroyer could have to unpeel several layers of the stack just to clear the
highlighted rectangle.

with this change, it automatically digs to the bottom or that stack and pulls
the weed out by the root instead. gone in one click.