swap out the whole element positioning scheme
and destroy the last locked target, not the click target

i don't remember why this didn't use `getClientBoundingRect` in the first
place, since it seems to work pretty well? is this old enough that that api
was still too new to rely on? :p

also adds a bit of transition flair that will probably get annoying.

targetting of fixed elements now works properly when the document is scrolled.